pyproject.toml
[tool.poetry]
name = "crypto"
version = "0.1.0"
description = ""
authors = ["Jesse Brink <jbrink0925@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.0"
bs4 = "^0.0.1"
requests = "^2.25.1"
dj-database-url = "^0.5.0"
django-allauth = "^0.44.0"
django-crispy-forms = "^1.11.1"
python-decouple = "^3.4"
psycopg2-binary = "^2.8.6"
requests-cache = "^0.5.2"
pytest-django = "^4.1.0"
gunicorn = "^20.0.4"
whitenoise = "^5.2.0"
lxml = "^4.6.2"
pytest-mypy = "^0.8.0"
pytest-vcr = "^1.0.2"
pytest-cov = "^2.11.1"
pycoingecko = "^1.4.0"
python-dateutil = "^2.8.1"
matplotlib = "^3.3.4"
pandas = "^1.2.3"
numpy = "^1.20.1"
image = "^1.5.33"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
